pyproject.toml
[tool.poetry]
name = "end to end q&a chatbot using google gemnii"
version = "0.1.0"
description = ""
authors = ["Nasir Abbas <akhtarabbas.islamicreaction@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
langchain-core = "^0.3.10"
langchain-google-genai = "^2.0.1"
python-dotenv = "^1.0.1"
streamlit = "^1.39.0"
langserve = "^0.3.0"
langchain-community = "^0.3.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
